# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface UploadUrlInput {
  name: string;
  size: number;
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface ResumableUploadResponse {
  /** Resumable session URL. The client PUTs chunks to this URL with a Content-Range header; an interrupted transfer resumes from the last committed byte instead of restarting. */
  sessionUrl: string;
  objectPath: string;
}

export interface Chapter {
  time: number;
  label: string;
}

export interface TranscriptSegment {
  start: number;
  end: number;
  text: string;
}

export type Visibility = typeof Visibility[keyof typeof Visibility];


export const Visibility = {
  private: 'private',
  public: 'public',
} as const;

/**
 * Upload lifecycle. "pending" rows exist from upload start; "ready" means the stored video was verified complete.
 */
export type RecordingStatus = typeof RecordingStatus[keyof typeof RecordingStatus];


export const RecordingStatus = {
  pending: 'pending',
  ready: 'ready',
} as const;

export type SelfieCorner = typeof SelfieCorner[keyof typeof SelfieCorner];


export const SelfieCorner = {
  'top-left': 'top-left',
  'top-right': 'top-right',
  'bottom-left': 'bottom-left',
  'bottom-right': 'bottom-right',
} as const;

export interface VisibilityInput {
  visibility: Visibility;
}

export interface RecordingInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  visibility?: Visibility;
  durationSec: number;
  trimStart: number;
  trimEnd: number;
  hasAudio?: boolean;
  videoPath: string;
  /** Byte size of the uploaded video blob. The server verifies the stored object matches this before persisting, so a truncated or failed upload can never produce a public link to incomplete media. */
  videoSize?: number;
  /** @nullable */
  thumbnailPath?: string | null;
  /** @nullable */
  gifPath?: string | null;
  selfieCorner?: SelfieCorner | null;
  chapters?: Chapter[];
  displayChaptersOnVideo?: boolean;
  notifyOnView?: boolean;
  transcript?: TranscriptSegment[];
}

export interface RecordingUpdateInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  durationSec: number;
  trimStart: number;
  trimEnd: number;
  hasAudio?: boolean;
  /** @nullable */
  gifPath?: string | null;
  selfieCorner?: SelfieCorner | null;
  chapters?: Chapter[];
  displayChaptersOnVideo?: boolean;
  notifyOnView?: boolean;
  transcript?: TranscriptSegment[];
}

export interface RecordingStartInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  visibility?: Visibility;
  durationSec: number;
  trimStart: number;
  trimEnd: number;
  hasAudio?: boolean;
  selfieCorner?: SelfieCorner | null;
  chapters?: Chapter[];
  displayChaptersOnVideo?: boolean;
  notifyOnView?: boolean;
}

export interface RecordingCompleteInput {
  /** @minLength 1 */
  videoPath: string;
  /** Byte size of the uploaded video blob; the server verifies the stored object matches before flipping the recording to ready. */
  videoSize?: number;
  /** @nullable */
  thumbnailPath?: string | null;
  /** @nullable */
  gifPath?: string | null;
}

export interface TranscriptInput {
  transcript: TranscriptSegment[];
}

export interface PlaybackUrl {
  /** Short-lived signed URL that streams directly from storage. */
  url: string;
  expiresInSec?: number;
}

/**
 * Lightweight listing shape for the library — everything a grid card needs, without the transcript/chapters payload.
 */
export interface RecordingSummary {
  shareId: string;
  title: string;
  description: string;
  visibility: Visibility;
  status: RecordingStatus;
  durationSec: number;
  trimStart: number;
  trimEnd: number;
  hasAudio: boolean;
  /** @nullable */
  thumbnailPath?: string | null;
  /** @nullable */
  gifPath?: string | null;
  selfieCorner?: SelfieCorner | null;
  views: number;
  createdAt: string;
}

export interface PublishedRecording {
  shareId: string;
  status: RecordingStatus;
  title: string;
  description: string;
  visibility: Visibility;
  durationSec: number;
  trimStart: number;
  trimEnd: number;
  hasAudio: boolean;
  videoPath: string;
  /** @nullable */
  thumbnailPath?: string | null;
  /** @nullable */
  gifPath?: string | null;
  selfieCorner?: SelfieCorner | null;
  chapters: Chapter[];
  displayChaptersOnVideo?: boolean;
  notifyOnView?: boolean;
  transcript: TranscriptSegment[];
  views: number;
  createdAt: string;
}

export interface ViewCount {
  views: number;
}

export interface VersionInfo {
  version: string;
  /** @nullable */
  releaseDate?: string | null;
}

export interface DesktopRelease {
  /** @nullable */
  version: string | null;
  /** @nullable */
  fileSize: number | null;
  notes: string;
  /** @nullable */
  downloadUrl: string | null;
  /** @nullable */
  updatedAt: string | null;
}

export interface DesktopReleaseInput {
  /** @minLength 1 */
  version: string;
  /** @minLength 1 */
  objectPath: string;
  fileSize: number;
  notes?: string;
}

export type AdminRole = typeof AdminRole[keyof typeof AdminRole];


export const AdminRole = {
  super_admin: 'super_admin',
  user: 'user',
} as const;

export interface UserGroup {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  memberCount: number;
  createdAt: string;
}

export interface AdminUser {
  id: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  role: AdminRole;
  groups: UserGroup[];
  recordingCount?: number;
  createdAt: string;
  /** @nullable */
  lastSignInAt?: string | null;
}

export interface RoleInput {
  role: AdminRole;
}

export interface GroupInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}

export interface GroupUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  description?: string | null;
}

export interface UserGroupAssignment {
  groupIds: number[];
}

export interface ImpersonationTicket {
  token: string;
  ticketUrl: string;
}

export type NotificationAudience = typeof NotificationAudience[keyof typeof NotificationAudience];


export const NotificationAudience = {
  all: 'all',
  group: 'group',
} as const;

export interface NotificationInput {
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  body: string;
  audience: NotificationAudience;
  /** @nullable */
  groupId?: number | null;
}

export interface NotificationResult {
  sent: number;
  failed: number;
  total: number;
  /** @nullable */
  message?: string | null;
}

export interface TosDocument {
  content: string;
  updatedAt: string;
}

export interface TosInput {
  /** @minLength 1 */
  content: string;
}

export type EmailPreviewSource = typeof EmailPreviewSource[keyof typeof EmailPreviewSource];


export const EmailPreviewSource = {
  our: 'our',
  clerk: 'clerk',
} as const;

export interface EmailPreview {
  id: string;
  name: string;
  description: string;
  subject: string;
  html: string;
  source: EmailPreviewSource;
}

export interface AdminRecordingActivity {
  shareId: string;
  title: string;
  views: number;
  visibility: Visibility;
  createdAt: string;
  /** @nullable */
  ownerEmail?: string | null;
}

export interface AdminSummary {
  totalUsers: number;
  totalRecordings: number;
  totalViews: number;
  recentSignups: AdminUser[];
  recentRecordings: AdminRecordingActivity[];
}