# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlInput"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
  /storage/uploads/resumable:
    post:
      operationId: requestResumableUpload
      tags: [storage]
      summary: Start a resumable upload session
      description: >-
        Create a resumable upload session in object storage for a large file.
        Returns a session URL the client uploads chunks to (PUT with
        Content-Range) so an interrupted transfer resumes from the last
        committed byte instead of restarting from zero.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlInput"
      responses:
        "200":
          description: Resumable upload session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumableUploadResponse"
...
components:
  schemas:
...
    UploadUrlInput:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
    ResumableUploadResponse:
      type: object
      required: [sessionUrl, objectPath]
      properties:
        sessionUrl:
          type: string
          description: >-
            Resumable session URL. The client PUTs chunks to this URL with a
            Content-Range header; an interrupted transfer resumes from the last
            committed byte instead of restarting.
        objectPath:
          type: string
...